requirements.txt
django==5.0.8
pyjwt==2.8.0
django-lti==0.6.0
django-csp==3.8
python-decouple==3.8
mysqlclient==2.2.4
django-mysql==4.14.0
gunicorn==22.0.0
whitenoise==6.7.0
debugpy==1.8.2
redis==5.0.7
canvasapi==3.3.0
djangorestframework==3.15.2
